requirements.txt
numpy==1.21.4
pandas==1.3.4
python-constraint==1.4.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
